% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDendro.R
\name{simpleDendro}
\alias{simpleDendro}
\title{Function to generate the clustering and dendrogram plots for the simpleHM function}
\usage{
simpleDendro(
  df,
  clust_samples = T,
  clust_params = T,
  force_symmetric_rows = F,
  param_order = NULL,
  linkage = "complete",
  pull_top = NULL,
  pull_side = NULL,
  normalise_params = T,
  normalise_samples = F,
  norm_method = "zscore",
  excluded_vars = c(),
  id_col = "",
  return_clustering = F
)
}
\arguments{
\item{df}{a dataframe to containing the data to be plotted. Required}

\item{clust_samples}{should the samples be clustered? T/F}

\item{clust_params}{should the variables be clustered? T/F}

\item{force_symmetric_rows}{should the rows be clustered same as the columns? T/F}

\item{param_order}{can be used to suplly a vector containg a custom order for the variables. Ignored if clust_params = TRUE}

\item{linkage}{which linkage method to use. defaults}

\item{pull_top}{should the dendrogram be rotated? if yes, what levels should be pulled to the front? Accepts numeric values (current order) and variable names}

\item{pull_side}{should the dendrogram be rotated? if yes, what levels should be pulled to the front? Accepts numeric values (current order) and variable names}

\item{normalise_params}{should the variables be normalized? defaults to T}

\item{normalise_samples}{should normalisation be applied to samples? defaults to F}

\item{norm_method}{which method should be applied to normalise? can be min/max scaling or z-score; T/F}

\item{excluded_vars}{character vector of variabels to be excluded}

\item{id_col}{a character with id column clearly identifying each sample}

\item{return_clustering}{should the clustering be returned? defaults to F}
}
\value{
returns a list of ggplot dendrogram plots. If return_clustering is set, returns the hclust objects instead
}
\description{
Function to generate the clustering and dendrogram plots for the simpleHM function
}
\examples{
df <- data.frame(samples = c(paste0("untreated", 1:6), paste0("treated", 7:12)),
                           group = c(rep("Untreated",6), rep("Treated",6)),
                           patient = c(rep(paste0("Patient", 1:3), 4)),
                           var1 = c(rnorm(6, 10, 1),  rnorm(6, 7, .7)),
                           var2 = c(rnorm(6, 10, 1),  rnorm(6, 200, 10)),
                           var3 = c(rnorm(6, 50, 5),  rnorm(6, 10, 1)),
                           var4 = c(rnorm(6, 10, 1),  rnorm(6, 60, .7)))

heatmap_dendro <- simpleDendro(df)
heatmap_dendro
}
