% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHM.R
\name{simpleHM}
\alias{simpleHM}
\title{Create a heatmap}
\usage{
simpleHM(
  df,
  clust_samples = T,
  clust_params = T,
  param_order = NULL,
  linkage = "complete",
  pull_top = NULL,
  pull_side = NULL,
  show_sample_names = T,
  normalise_params = T,
  normalise_samples = F,
  norm_method = "zscore",
  excluded_vars = c(),
  id_col = "",
  color_code = c(high = "#FF1c00", low = "darkblue"),
  custom_threshold = NULL,
  outlier.removal = T,
  outlier.threshold = 0.95,
  add_annotation = F,
  anno_col = "",
  annotation_colors = list(),
  add_dendros = F,
  .plot = T,
  return_list = F
)
}
\arguments{
\item{df}{a dataframe to containing the data to be plotted. Required}

\item{clust_samples}{should the samples be clustered? T/F}

\item{clust_params}{should the variables be clustered? T/F}

\item{param_order}{can be used to suplly a vector containg a custom order for the variables. Ignored if clust_params = TRUE}

\item{linkage}{which linkage method to use. defaults}

\item{pull_top}{should the dendrogram be rotated? if yes, what levels should be pulled to the front? Accepts numeric values (current order) and variable names}

\item{pull_side}{should the dendrogram be rotated? if yes, what levels should be pulled to the front? Accepts numeric values (current order) and variable names}

\item{show_sample_names}{should the sample names be shown in the heatmap?}

\item{normalise_params}{should the variables be normalized? defaults to T}

\item{normalise_samples}{should normalisation be applied to samples? defaults to F}

\item{norm_method}{which method should be applied to normalise? can be min/max scaling or z-score; T/F}

\item{excluded_vars}{character vector of variabels to be excluded}

\item{id_col}{a character with id column clearly identifying each sample}

\item{color_code}{color code for the fill aestethic. default is "#FF1c00" for high values and "darkblue" for low values}

\item{custom_threshold}{custom threshold for the scale; defaults to NULL}

\item{outlier.removal}{should outlier be removed? i.e. single values which drive the normalized values; default is T}

\item{outlier.threshold}{threshold for outlier.removal. currently is set to the 95th percentile (0.95)}

\item{add_annotation}{should annotation be added to the heatmap? default = F}

\item{anno_col}{which columns is the data contain annotation? can be a character vector.}

\item{annotation_colors}{a list containing the vectors with colors for every annotation column}

\item{add_dendros}{should dendros be added}

\item{.plot}{should the plot be created and returned? default is T, if F just returns the data}

\item{return_list}{when adding annotation, should the assembled plot be returned? or should the components be returned as a list? defaults to F}
}
\value{
returns a heatmap as a ggplot object. If annotation is added, a patchwork assembled plot is returned.
}
\description{
Create a heatmap
}
\examples{
df <- data.frame(samples = c(paste0("untreated", 1:6), paste0("treated", 7:12)),
                           group = c(rep("Untreated",6), rep("Treated",6)),
                           patient = c(rep(paste0("Patient", 1:3), 4)),
                           var1 = c(rnorm(6, 10, 1),  rnorm(6, 7, .7)),
                           var2 = c(rnorm(6, 10, 1),  rnorm(6, 200, 10)),
                           var3 = c(rnorm(6, 50, 5),  rnorm(6, 10, 1)),
                           var4 = c(rnorm(6, 10, 1),  rnorm(6, 60, .7)))

heatmap_plot <- simpleHM(df)

anno_added <- simpleHM(df, add_annotation = TRUE)

}
